move progress components up a level in docs
