+++ packages/core/src/components/_index.scss
@@ -20,7 +20,7 @@
 @import "overlay/overlay";
 @import "popover/popover";
 @import "portal/portal";
-@import "progress/progress-bar";
+@import "progress-bar/progress-bar";
 @import "skeleton/skeleton";
 @import "slider/slider";
 @import "spinner/spinner";
