+++ packages/core/src/components/index.ts
@@ -49,7 +49,7 @@ export * from "./overlay/overlay";
 export * from "./text/text";
 export * from "./popover/popover";
 export * from "./portal/portal";
-export * from "./progress/progressBar";
+export * from "./progress-bar/progressBar";
 export * from "./slider/rangeSlider";
 export * from "./slider/slider";
 export * from "./spinner/spinner";
