+++ packages/core/src/components/progress-bar/_common.scss
