+++ packages/core/src/components/progress-bar/progress-bar.md
