+++ packages/core/src/components/progress-bar/progressBar.tsx
