--- packages/core/src/components/progress/progress.md
@@ -1,7 +1,0 @@
-@# Progress
-
-Blueprint provides two ways to indicate progress: a horizontal progress bar and a circular spinner.
-
-@page progress-bar
-@page spinner
-@page skeleton
