+++ packages/core/src/components/spinner/_spinner.scss
@@ -2,7 +2,7 @@
 // Licensed under the terms of the LICENSE file distributed with this project.
 
 @import "../../common/variables";
-@import "../progress/common";
+@import "../progress-bar/common";
 @import "./common";
 
 /*
